Add option to wait for GUI session
Wait for GUI agent to report that GUI daemon has connected. This in
practice waits for the user to log into GUI domain (or dom0).
This is relevant if the qrexec policy for qubes.Input* is set to ask -
if the user is not logged in yet, the qrexec prompt cannot be displayed,
and the qrexec prompt would be denied.

Keep the change compatible with the old GUI agent (the same package may
be installed on R4.2) - if the status file is not there, skip the wait.
And to make the update least intrusive, only recommend pyinotify, but if
it's not there, fallback to polling the file.

QubesOS/qubes-issues#9940
https://github.com/QubesOS/qubes-gui-agent-linux/pull/251#issuecomment-3649371068